WIP website navigation
I've made the basic shapes for most of the parts of the website and am focusing on the navigation area, formerly the header. Looking to make this work before working on the pages, which will likely have few components to them. After this I'll go onto more UI features and accessibility